[WIP] Add option to summarise functions
Summarised functions are not called. Instead, their return values are
symbolic, and any pointer passed to them is written into with symbolic
values until the end of its capacity.

Current issue: If a pointer is illegal (especially if it is NULL), the
call fails. However, it should be possible for the call to continue
without modifying the 'invalid' memory address.